refactor(frontend): extract self-contained hooks from agent-view.tsx
