+++ .changesets/1782957522-refactor-frontend-extract-hooks-from-agent-view-if7v.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(frontend): extract hooks from agent-view
